Deal with a new form of brain-damaged message-IDs
This might be a Gmail problem---I'm 100% not sure yet---but
Message-ids of the form <some-string>@<some-IP-address> are supposed
to have the IP address enclosed in square brackets.  Unfortunately,
sometimes they are omitted, which screws up mailsync's ability to
identify messages as identical on both sides of the connection.  So
now we add the square brackets where they are missing.

The check for patterns that look like IP addresses could be less hacky
and more precise.  It might be worth rewriting at some point